Fixed the LogScale class's ticks() function to correctly handle the case where the length of the interval is less than a decade.  It will now not generate ticks outside the given limits.
Fixed another LogScale edge-case bug that could occur when the steps between ticks was tenths of decades; in some cases, the left-most or right-most tick could be repeated.

Added more unit tests, and split some existing tests into multiple tests.

Fixed a test case that was failing when it shouldn't.

NOTE: There is still one unit test failing in scales/scales_test_case.py.  I'm not sure how this is supposed to work, so I haven't tried to fix it.