TST: add test for bug in setting xbounds and ybounds in countour and image plots
Commit 1e2431fcd42684676163acf6b275c1ef4d5fd87b modified the logic of
interpreting xbounds and ybounds to create contour and image plots.

In the case xbounds or ybounds are 2D array, it introduces a bug, as
the shape of the final grid is too small.

This case arises when doing something like

x,y = meshgrid(xs,ys)
z = sin(x)*x*jn(0,y)
contour(x,y,z, bgcolor="black")

as in the shell/contour.py example (which is broken)