TST: logmapper and linearmapper have different behaviors when passed a scalar
the first returns a 1D array, the second a scalar. This causes problems
in code that relies on one or the other. The commit adds two tests
that check for a consistent behavior.

making the tests pass would be trivial, but at present I'm not sure
how much of the code relies on one or the other standard

the recommended workaround is to *always* pass the mappers a 1D array