FIX: origin in jitterplot depends on orientation
Since JitterPlot is a 1D plot, it overwrote ``origin`` to specify it as one of the edges (e.g., 'left', or 'top', instead of 'bottom right').

However, for instance, origin=='bottom' does not make sense for an horizontal JitterPlot, and the implementation ended up creating a plot where the horizontal width depended on the vertical size of the window.

In the new implementation, origin is as usual, and the code takes into account  only the horizontal or vertical part of it, depending on the orientation of the plot.