Changes to image plots to allow speedups for large images
This adds code which allows:
  * colourmaps to be applied only to currently visible region of an image plot
  * downsampling of images
These features are optional, default behaviour is unchanged.

The trade-off is increased computation on every change, vs slower updates and
higher memory usage when the cache does change for the current code.  This is
intended to be used when Chaco is dealing with large image plots which may be
rapidly changing in reaction to user interaction (eg. selecting a region of
an image plot).