Remove wx-dependency from data_stream example.
Reuse generic Traits classes instead of requiring a wx app. This change was
thoroughly tested in ipython (with all permutations of wx/qt backends and
ETS_TOOLKITs). This was run in Mac and Windows version of the IDE, and
I checked demo.py with ETS_TOOLKIT set to wx and qt.