Update multi_colored_scatter.py
The previous method of initializing the numpy array of color tuples was incorrect. The result was an array of records which was suitable for rendering to screen but caused a problem when outputting to PDF. The cause of this problem is the technique used to slice the tuple to check for the existence of the alpha value. Initializing the array as an array of float64 tuples has solved the problem.