A couple of fixes to keep VS2010 compatibility and avoid a deprecated warnings with newer versions of numpy
The first fix is on chaco\contour\cntrc.c to move the moduledef
declaration to the top the MOD_INIT function to be complain with older
compilers followingn the old C89 specification (such as VS 2010), the
credit to spot this goes to @stevenengler . The change should be
compatible with any other compiler.

The second change is on chaco\base.py to avoid a deprecate warning on
newer versions of numpy that do not accept differantion of boolean
arrays. The current fix simply recasted the array as uint8.