REF : Use enable Container directly for chaco plot containers
instead of relying on BasePlotContainer which itself inherits from
Container.

We can do so now because we moved all of relevant traits down to the
subclasses from the BasePlotContainer in the previous commit. Note that
there are still a few traits on the BasePlotContainer but those are
deprecated and are being removed elsewhere

	modified:   chaco/plot_containers.py