REF : Use enable utils in chaco StackedPlotContainer
specifically, in the methods StackedPlotContainer.get_preferred_size and
StackedPlotContainer._do_layout.

note that the get_preferred_size method is slightly non-trivial because
the underlying enable util stacked_preferred_size does not fully cover
the needs of the method - we might want to fix the underlying enable
util to do so.

	modified:   chaco/plot_containers.py