REF : Chaco VPlotContainer now inherits from enable VStackedContainer
and we removed a bunch of traits from the subclass which are already
defined in the baseclass.

ideally, we would also be able to remove the get_preferred_size and
persistence (__getstate__) methods from the subclass but once those are
fixed in enable and a new version of enable is available, we can clean
this code up further.

	modified:   chaco/plot_containers.py