Add Python 3.8 tests (#851)
This adds a parallel set of tests for Python 3.8 rather than trying to adapt the current set of tests.  Also updates edmtool to allow Python 3.8 and PySide6, and pyproject.toml to allow for known issue with setuptools.

Needed to make one change to code due to newer version of NumPy being picked up.

No requirement that they pass for PRs to be mergeable yet, but will likely add that once this PR is approved.