MAINT: fix DemoController in the image_from_file example (#893)
After the update of traitsui, the ui requires the init() method to return a boolean that indicates whether the initiation has been successful. However, some demos have not been able to adapt to this change such as #889 . This PR aims to fix #889 by adding the boolean return to the demo. closes #889 

Co-authored-by: Chengyu Liu <cyliu@aus552cyliu.local>
Co-authored-by: Corran Webster <cwebster@enthought.com>