removed typo by replacing `obseve()` with `observe()`, fixing issue #914 (#915)
As this fixes a typo clearly visible in the source code, I have not
added any code to reproduce the error nor a unit test. Hope this is
still fine (I'm short on time currently, sorry!).

See [Issue 914](https://github.com/enthought/chaco/issues/914) and/or
the original pull request #590 that introduced this bug.

Co-authored-by: Ioannis Tziakos <ioannist@enthought.com>